// server/researchindicators/src/domain/entities/green-checks/dto/message-oicr.dto.ts
export class MessageOicrDto {
  requester_by: string;
  title: string;
  reviewed_by: string;
  decision_date: string;
  justification: string;
  url: string;
  oicr_number: string;
  mel_expert_name: string;
  sharepoint_url: string;
  requester_by_email: string;
  reviewed_by_email: string;
  mel_expert_email: string;
}